requirements.txt
absl-py==2.1.0
altair==5.3.0
astunparse==1.6.3
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
flatbuffers==24.3.25
fonttools==4.53.0
gast==0.5.4
gitdb==4.0.11
gitpython==3.1.43
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
humanfriendly==10.0
idna==3.7
jinja2==3.1.4
jsonschema-specifications==2023.12.1
jsonschema==4.22.0
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
markdown-it-py==3.0.0
markdown==3.6
markupsafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
numpy==1.26.4
onnx==1.16.1
onnxruntime==1.18.1
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
pyarrow==16.1.0
pydeck==0.9.1
pygments==2.18.0
pyparsing==3.1.2
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
scipy==1.14.0
setuptools==70.0.0
six==1.16.0
smmap==5.0.1
streamlit-drawable-canvas==0.9.3
streamlit==1.35.0
sympy==1.12.1
tenacity==8.3.0
tensorboard-data-server==0.7.2
tensorboard==2.16.2
tensorflow==2.16.1
termcolor==2.4.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.1
werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
